requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg==3.1.9
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
